Split phone number form selectors into their own file
