+++ app/javascript/selectors/screening/peopleFormSelectors.js
@@ -5,7 +5,6 @@ import {getSSNErrors} from 'utils/ssnValidator'
 import {isRequiredIfCreate, combineCompact} from 'utils/validator'
 import moment from 'moment'
 import {getScreeningIdValueSelector} from 'selectors/screeningSelectors'
-import PHONE_NUMBER_TYPE from 'enums/PhoneNumberType'
 import {RACE_DETAILS} from 'enums/Races'
 
 export const getPeopleSelector = (state) => state.get('peopleForm')
@@ -210,16 +209,6 @@ export const getPersonAlertErrorMessageSelector = (state, personId) => {
   }
   return undefined
 }
-
-export const getPhoneNumberTypeOptions = () => fromJS(PHONE_NUMBER_TYPE.map((type) => ({value: type, label: type})))
-export const getPersonPhoneNumbersSelector = (state, personId) => (
-  state.get('peopleForm', Map()).get(personId).get('phone_numbers', List()).map((phoneNumber) => (
-    Map({
-      number: phoneNumber.getIn(['number', 'value']) || '',
-      type: phoneNumber.getIn(['type', 'value']) || '',
-    })
-  ))
-)
 
 export const getFilteredPersonRolesSelector = (state, personId) => {
   const selectedRoles = state.getIn(['peopleForm', personId, 'roles', 'value'], List())
